chore: dedicated cmd utils
